Bring back only actual possible implementation of `remove_matches`
Literally had it in my head (and why, I don't know) for like six hours that CharSearcher pre-collected
the match tuples into some kind of internal storage. It does not, and is a lazy iterator because of
course it is, meaning you do indeed have to place the match tuples in your own intermediate storage
in any scenario where you want to grab them all ahead of time.